Move Link component to separate folder
