+++ desktop/packages/mullvad-vpn/src/renderer/lib/components/typography/index.ts
@@ -6,5 +6,5 @@ export * from './Text';
 export * from './TitleBig';
 export * from './TitleLarge';
 export * from './TitleMedium';
-export * from './Link';
+export * from '../link/Link';
 export * from './Label';
